Refactor: 유저 컨트롤러, 기업 컨트롤러 분리
