+++ bin/main/mapper/apply.xml
@@ -13,7 +13,7 @@
       select * from apply_tb
    </select>
 
-   <select id="findOne" resultType="shop.mtcoding.job.model.apply.Apply">
+   <select id="findById" resultType="shop.mtcoding.job.model.apply.Apply">
       select * from apply_tb where id = #{id}  
    </select>
