+++ bin/main/mapper/like.xml
@@ -13,7 +13,7 @@
       select * from like_tb
    </select>
 
-   <select id="findOne" resultType="shop.mtcoding.job.model.like.Like">
+   <select id="findById" resultType="shop.mtcoding.job.model.like.Like">
       select * from like_tb where id = #{id}  
    </select>
