+++ bin/main/mapper/recruitment_post.xml
@@ -14,7 +14,7 @@
       select * from resume_tb
    </select>
 
-   <select id="findOne" resultType="shop.mtcoding.job.model.recruitmentPost.RecruitmentPost">
+   <select id="findById" resultType="shop.mtcoding.job.model.recruitmentPost.RecruitmentPost">
       select * from resume_tb where id = #{id}  
    </select>
