+++ bin/main/mapper/resume.xml
@@ -16,7 +16,7 @@
       select * from apply_tb
    </select>
 
-   <select id="findOne" resultType="shop.mtcoding.job.model.resume.Resume">
+   <select id="findById" resultType="shop.mtcoding.job.model.resume.Resume">
       select * from apply_tb where id = #{id}  
    </select>
