+++ src/main/java/shop/mtcoding/job/controller/UserController.java
@@ -6,9 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 
-import shop.mtcoding.job.dto.user.UserReqDto.JoinEnterpriseReqDto;
 import shop.mtcoding.job.dto.user.UserReqDto.JoinUserReqDto;
-import shop.mtcoding.job.dto.user.UserReqDto.LoginEnterpriseReqDto;
 import shop.mtcoding.job.dto.user.UserReqDto.LoginUserReqDto;
 import shop.mtcoding.job.handler.exception.CustomException;
 import shop.mtcoding.job.model.user.User;
@@ -23,7 +21,7 @@ public class UserController {
 
     @GetMapping("/loginForm")
     public String loginForm() {
-        return "user/loginForm";
+        return "login/loginForm";
     }
 
     @PostMapping("/user/login")
@@ -48,28 +46,6 @@ public class UserController {
         return "redirect:/";
     }
 
-    @PostMapping("/enterprise/login")
-    public String enterpriseLogin(LoginEnterpriseReqDto loginEnterpriseReqDto) {
-        if (loginEnterpriseReqDto.getEnterpriseName() == null || loginEnterpriseReqDto.getEnterpriseName().isEmpty()) {
-            throw new CustomException("아이디를 작성해주세요");
-        }
-        if (loginEnterpriseReqDto.getPassword() == null || loginEnterpriseReqDto.getPassword().isEmpty()) {
-            throw new CustomException("비밀번호를 작성해주세요");
-        }
-        // 1. 로그인하기 service
-        User principal = userService.기업로그인하기(loginEnterpriseReqDto);
-
-        // 2. session에 저장
-        session.setAttribute("principal", principal);
-
-        // 3. principal 유효성 검사
-        if (session.getAttribute("principal") == null) {
-            throw new CustomException("존재하지 않는 아이디거나 비밀번호를 다시 확인해주시기 바랍니다");
-        }
-
-        return "redirect:/";
-    }
-
     @GetMapping("/logout")
     public String logout() {
         session.invalidate();
@@ -78,7 +54,7 @@ public class UserController {
 
     @GetMapping("/joinForm")
     public String joinForm() {
-        return "user/joinForm";
+        return "join/joinForm";
     }
 
     @PostMapping("/user/join")
@@ -100,36 +76,6 @@ public class UserController {
         }
 
         userService.유저가입하기(joinUserReqDto);
-
-        return "redirect:/loginForm";
-    }
-
-    @PostMapping("/enterprise/join")
-    public String enterpriseJoin(JoinEnterpriseReqDto joinEnterpriseReqDto) {
-
-        if (joinEnterpriseReqDto.getEnterpriseName() == null || joinEnterpriseReqDto.getEnterpriseName().isEmpty()) {
-            throw new CustomException("enterprisename을 작성해주세요");
-        }
-        if (joinEnterpriseReqDto.getPassword() == null || joinEnterpriseReqDto.getPassword().isEmpty()) {
-            throw new CustomException("password를 작성해주세요");
-        }
-        if (joinEnterpriseReqDto.getAddress() == null || joinEnterpriseReqDto.getAddress().isEmpty()) {
-            throw new CustomException("address를 작성해주세요");
-        }
-        if (joinEnterpriseReqDto.getEmail() == null || joinEnterpriseReqDto.getEmail().isEmpty()) {
-            throw new CustomException("email을 작성해주세요");
-        }
-        if (joinEnterpriseReqDto.getContact() == null || joinEnterpriseReqDto.getContact().isEmpty()) {
-            throw new CustomException("전화번호를 입력해주세요");
-        }
-        if (joinEnterpriseReqDto.getSector() == null || joinEnterpriseReqDto.getSector().isEmpty()) {
-            throw new CustomException("sector을 작성해주세요");
-        }
-        if (joinEnterpriseReqDto.getSize() == null || joinEnterpriseReqDto.getSize().isEmpty()) {
-            throw new CustomException("size을 작성해주세요");
-        }
-
-        userService.기업가입하기(joinEnterpriseReqDto);
 
         return "redirect:/loginForm";
     }
