+++ src/main/java/shop/mtcoding/job/model/user/UserRepository.java
@@ -23,4 +23,7 @@ public interface UserRepository {
         public User findByUsernameAndPassword(@Param("username") String username, @Param("password") String password);
 
         public String findSaltById(int id);
-        }
+
+        public User findByName(String username);
+
+}
