+++ src/main/java/shop/mtcoding/job/service/EnterpriseService.java
@@ -7,27 +7,19 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import shop.mtcoding.job.dto.user.UserReqDto.JoinEnterpriseReqDto;
-import shop.mtcoding.job.dto.user.UserReqDto.JoinUserReqDto;
 import shop.mtcoding.job.dto.user.UserReqDto.LoginEnterpriseReqDto;
-import shop.mtcoding.job.dto.user.UserReqDto.LoginUserReqDto;
 import shop.mtcoding.job.handler.exception.CustomException;
 import shop.mtcoding.job.model.enterprise.EnterpriseRepository;
 import shop.mtcoding.job.model.user.User;
-import shop.mtcoding.job.model.user.UserRepository;
-import shop.mtcoding.job.util.HashEncoding;
-import shop.mtcoding.job.util.saltEncoder;
-import shop.mtcoding.job.util.sha256Encoder;
 
 @Service
-public class UserService {
-    @Autowired
-    private UserRepository userRepository;
+public class EnterpriseService {
 
     @Autowired
     private EnterpriseRepository enterpriseRepository;
 
     @Transactional(readOnly = true)
-    public User 유저로그인하기(LoginUserReqDto loginUserReqDto) {
+    public User 기업로그인하기(LoginEnterpriseReqDto loginEnterpriseReqDto) {
         try {
             String sha256Hash = sha256Encoder.sha256(loginUserReqDto.getPassword());
             String salt = saltEncoder.salt();
@@ -41,16 +33,15 @@ public class UserService {
     }
 
     @Transactional
-    public void 유저가입하기(JoinUserReqDto joinUserReqDto) {
+    public void 기업가입하기(JoinEnterpriseReqDto joinEnterpriseReqDto) {
         // 1. 유저 유효성 검사
-        User sameuser = userRepository.findByName(joinUserReqDto.getUsername());
+        User sameuser = userRepository.findByName(joinEnterpriseReqDto.getUsername());
 
         if (sameuser != null) {
             throw new CustomException("동일한 아이디가 존재합니다");
         }
         // 1. db에 insert하기
-        int result = userRepository.insert(joinUserReqDto.getUsername(), joinUserReqDto.getPasswor
-                (), joinUserReqDto.getEmail());
+        int result = userRepository.insert(joinReqDto.getUsername(), joinReqDto.getPassword(), joinReqDto.getEmail());
 
         if (result != 1) {
             throw new CustomException("회원가입이 실패하였습니다");
