+++ src/main/resources/mapper/user.xml
@@ -14,7 +14,7 @@
       select * from user_tb
    </select>
 
-   <select id="findOne" resultType="shop.mtcoding.job.model.user.User">
+   <select id="findById" resultType="shop.mtcoding.job.model.user.User">
       select * from user_tb where id = #{id}  
    </select>
 
@@ -28,5 +28,9 @@
 
    <select id="findSaltById" resultType="shop.mtcoding.job.model.user.User">
       select salt = #{salt} from user_tb where id = #{id}  
+   </select>
+
+   <select id="findByName" resultType="shop.mtcoding.job.model.user.User">
+      select * from user_tb where username = #{username}  
    </select>
 </mapper>
