+++ src/main/webapp/WEB-INF/view/join/joinForm.jsp
