+++ src/main/webapp/WEB-INF/view/layout/header.jsp
@@ -55,7 +55,7 @@
                                         <a class="nav-link text-dark" href="/loginForm">로그인</a>
                                     </li>
                                     <li class="nav-item">
-                                        <a class="nav-link text-dark" href="/join">회원가입</a>
+                                        <a class="nav-link text-dark" href="/joinForm">회원가입</a>
                                     </li>
                                     </c:when>
